// src/organizations/models/organization.generated.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file.
 *
 * =>  @see ./src/schemas/organizations/Organization.yml
 *
 * and run "npm run schema-to-interface" to regenerate this file.
 */

export interface Organization {
  identifier: string;
  name: string;
  description?: string;
  createdAt?: string;
  updatedAt?: string;
}